Add combat UI tabs
